==> 06_IsolateAnchorsFromVideo.ipynb <==
import numpy as np
import pandas as pd

import cv2
import os

base_dir = '/Users/vivekhsridhar/Library/Mobile Documents/com~apple~CloudDocs/Documents/Code/Python/SpatialRegistration/io/'
output_dir = '/Users/vivekhsridhar/Library/Mobile Documents/com~apple~CloudDocs/Documents/Metashape/TalChhapar/images/'

df = pd.read_csv(base_dir + '20230310_SE_Lek1_P1D1_DJI_0190_Anchored.csv')
df['best_anchor_frame'] = df['best_anchor_frame'].astype(int)
df.head()

# Step 1: Extract unique frame numbers from the 'anchor_frames' column
unique_frames = df['best_anchor_frame'].unique()

# Define the video path
video_path = base_dir + '20230310_SE_Lek1_P1D1_DJI_0190.MP4'  # Specify the correct video path
video_name = os.path.splitext(os.path.basename(video_path))[0]  # Extract video name without extension

# Step 2: Create a directory with the video name to store extracted frames
output_folder = output_dir + f'{video_name}'
os.makedirs(output_folder, exist_ok=True)

# Step 3: Use OpenCV to open the video and extract frames
cap = cv2.VideoCapture(video_path)

# Get the total number of frames in the video
total_frames = int(cap.get(cv2.CAP_PROP_FRAME_COUNT))

# Check if the video opened successfully
if not cap.isOpened():
    print("Error: Could not open the video.")
else:
    print(f"Total frames in video: {total_frames}")
    
    # Step 4: Iterate through unique frames, read and save them
    for frame_no in unique_frames:
        # Ensure frame number is within the video frame range
        if frame_no >= 0 and frame_no < total_frames:
            # Set the video frame to the specific frame number
            cap.set(cv2.CAP_PROP_POS_FRAMES, frame_no)
            
            # Read the frame
            ret, frame = cap.read()
            
            if ret:
                # Create the filename for the frame
                frame_filename = os.path.join(output_folder, f"{video_name}_frame{frame_no}.jpg")
                
                # Save the frame as an image
                cv2.imwrite(frame_filename, frame)
                print(f"Saved frame {frame_no}")
            else:
                print(f"Error: Could not read frame {frame_no}")
        else:
            print(f"Frame number {frame_no} is out of bounds")

# Release the video capture object
cap.release()
