==> 06_TransformToAnchorCoordinates.ipynb <==
import numpy as np
import pandas as pd
import matplotlib.pyplot as plt
import seaborn as sns

import os
import glob
import pickle
import helper_functions_06b as hf

DATE = ['20230311']#, '20230305']
SESSION = ['SE_Lek1']#, 'SE_Lek1']
DRONE = ['P1D1']#, 'P2D4', 'P3D5', 'P3D6']

tracking_directory = '/Volumes/EAS_shared/blackbuck/working/processed/Field_Recording_2023/SpatialRegistration'

# Thresholds for translation, rotation, and scale
thresholds = {
    'T_translation': 30,  # e.g., 30 pixels
    'T_rotation': 20,     # e.g., 20 degrees
    'T_scale': 0.1         # e.g., 10% change
}

for date in DATE:
    for session in SESSION:
        for drone in DRONE:
            
            internal_path = tracking_directory + '/' + date + '/' + session + '/' + drone
            files = glob.glob(internal_path + '/' + date + '_' + session + '_' + drone + '*_tracked.csv')
            
            for file in sorted(files):
                print(date, session, drone, os.path.basename(file))
                
                df = pd.read_csv(file)
                df['x'] = df['bb_left'] + df['bb_width']/2
                df['y'] = df['bb_top'] + df['bb_height']/2
                df['idx'] = df['track_id']

                df = df.loc[:,['frame', 'x', 'y', 'idx']]
                
                # Step 1: Identify anchor frames
                anchor_frames = hf.identify_anchor_frames(df, thresholds)

                # Step 2: Find the best anchor for each frame
                df = hf.find_best_anchor_for_each_frame_parallel(df, anchor_frames)

                # Step 3: Apply homography using the best anchor for each frame
                df, homography_matrices, rms_errors = hf.apply_homography_to_best_anchor(df)
                
                anchor_filename = os.path.basename(file).replace('_YOLO_tracked.csv', '_Anchored.csv')
                pickle_filename = os.path.basename(file).replace('_YOLO_tracked.csv', '_homographies.pkl')
                df.to_csv(os.path.join(os.path.dirname(file), anchor_filename), index=False, mode='w')
                with open(os.path.join(os.path.dirname(file), pickle_filename), 'wb') as f:
                    pickle.dump(homography_matrices, f)
