==> 07_IsolateAnchorFrames.ipynb <==
import numpy as np
import pandas as pd

import cv2
import os
import glob
import helper_functions_07b as hf

from multiprocessing import Pool

DATE = ['20230308']#, '20230306']
SESSION = ['SE_Lek1']#, 'SE_Lek1']
DRONE = ['P3D6']#, 'P1D2', 'P2D3', 'P2D4', 'P3D5', 'P3D6']

files_directory = '/Volumes/EAS_shared/blackbuck/working/processed/Field_Recording_2023/SpatialRegistration'
video_directory = '/Volumes/EAS_shared/blackbuck/working/rawdata/Field_Recording_2023/Original/lekking'

# Generate all tasks
tasks = [(date, session, drone, files_directory, video_directory) for date in DATE for session in SESSION for drone in DRONE]

# Use multiprocessing Pool to parallelize
with Pool(processes=6) as pool:  # Adjust the number of processes based on the number of tasks or on your CPU
    pool.map(hf.process_task, tasks)
