==> 10_TerritoryIDMatch.ipynb <==
import subprocess
import glob

DATE = ['20230318']#, '20230314', '20230316']
SESSION = ['SM_Lek1']#, 'SE_Lek1']

files_directory = '/Volumes/SSD5'

for date in DATE:
    for session in SESSION:
        print(date, session)

        internal_path = files_directory + '/' + date + '/' + session + '/'
        metashape_csv = glob.glob(internal_path + '/TalChhapar_' + date + '_' + session + '/output/' + '*.csv')

        dlc_folders = [
#             files_directory + '/' + date + '/' + session + '/' + 'P1D1',
#             files_directory + '/' + date + '/' + session + '/' + 'P1D2',
#             files_directory + '/' + date + '/' + session + '/' + 'P2D3',
#             files_directory + '/' + date + '/' + session + '/' + 'P2D4',
            files_directory + '/' + date + '/' + session + '/' + 'P3D5',
#             files_directory + '/' + date + '/' + session + '/' + 'P3D6'
        ]

        for folder in dlc_folders:
            subprocess.run([
                "python3", "batch_runner_10b.py",
                '-m', metashape_csv[0],
                '-d', folder
            ])
