==> 15_ReformatAnchoredTrajectories.ipynb <==
import pandas as pd
import matplotlib.pyplot as plt

import os
import glob

DATE = ['20230313']#, '20230309', '20230311', '20230312']
SESSION = ['SE_Lek1']#, 'SE_Lek1']
DRONE = ['P1D1', 'P1D2', 'P2D3', 'P2D4', 'P3D5', 'P3D6']

files_directory = '/Volumes/SSD5'

for date in DATE:
    for session in SESSION:
        for drone in DRONE:
            print(date, session, drone)
            
            internal_path = files_directory + '/' + date + '/' + session + '/' + drone
            files = glob.glob(internal_path + '/' + date + '_' + session + '_' + drone + '*_Anchored_trajectories.csv')
            
            for file in sorted(files): 
                df = pd.read_csv(file)
                df.columns = ['frame', 'x', 'y', 'Point', 'class_id', 'class_name', 'best_anchor_frame', 'u', 'v']

                # Load the corresponding matched file
                matched_file = file.replace('_Anchored_trajectories.csv', '_Anchored_matched.csv')
                if os.path.exists(matched_file):
                    matched_df = pd.read_csv(matched_file)

                    # Filter df to only include frames present in matched_df
                    df = df[df['frame'].isin(matched_df['frame'])].copy()

                    filename = os.path.basename(file).rsplit('_Anchored_trajectories', 1)[0]
                    df['Camera'] = filename + '_frame' + df['best_anchor_frame'].astype(int).astype(str)
                    df['video'] = filename

                    out_filename = os.path.basename(file).replace('_Anchored_trajectories.csv', '_trajectories_uv.csv')
                    df.to_csv(os.path.join(os.path.dirname(file), out_filename), index=False, mode='w')
                else:
                    print(f"Matched file not found: {matched_file}")
