==> 17_ConvertUTM.ipynb <==
import pandas as pd
import utm

import os
import glob

DATE = ['20230314']#, '20230309']
SESSION = ['SE_Lek1']#, 'SE_Lek1']
DRONE = ['P1D1', 'P1D2', 'P2D3', 'P2D4', 'P3D5', 'P3D6']

files_directory = '/Volumes/SSD5'

for date in DATE:
    for session in SESSION:
        for drone in DRONE:
            print(date, session, drone)
            
            internal_path = files_directory + '/' + date + '/' + session + '/' + drone
            files = glob.glob(internal_path + '/' + date + '_' + session + '_' + drone + '*_3D_territories.csv')
            
            for file in sorted(files): 
                df = pd.read_csv(file)
                territories_utm = utm.from_latlon(df.loc[:,'latitude'].values, df.loc[:,'longitude'].values)
                df['easting'] = territories_utm[0]
                df['northing'] = territories_utm[1]
                
                filename = os.path.basename(file).replace('_3D_territories.csv', '_3D_territories_utm.csv')
                df.to_csv(os.path.join(os.path.dirname(file), filename), index=False, mode='w')

for date in DATE:
    for session in SESSION:
        for drone in DRONE:
            print(date, session, drone)
            
            internal_path = files_directory + '/' + date + '/' + session + '/' + drone
            files = glob.glob(internal_path + '/' + date + '_' + session + '_' + drone + '*_3D_trajectories.csv')
            
            for file in sorted(files): 
                df = pd.read_csv(file)
                territories_utm = utm.from_latlon(df.loc[:,'latitude'].values, df.loc[:,'longitude'].values)
                df['easting'] = territories_utm[0]
                df['northing'] = territories_utm[1]
                
                filename = os.path.basename(file).replace('_3D_trajectories.csv', '_3D_trajectories_utm.csv')
                df.to_csv(os.path.join(os.path.dirname(file), filename), index=False, mode='w')
